browser_session_controller/requirements.txt
fastapi
uvicorn[standard]
httpx
supabase
PyJWT
kubernetes

